Some refactorings, further work on the ServiceNow sink
