+++ src/robusta/integrations/mail/sender.py
@@ -9,51 +9,23 @@ from apprise.AppriseAttachment import AppriseAttachment
 
 from robusta.core.reporting.base import BaseBlock, Emojis, Finding, FindingStatus
 from robusta.core.reporting.blocks import (
-    FileBlock,
     LinksBlock,
     LinkProp,
     MarkdownBlock,
 )
 from robusta.core.reporting.consts import EnrichmentAnnotation, FindingSource
+from robusta.core.sinks.common.html_tools import HTMLBaseSender, HTMLTransformer, with_attr
 from robusta.core.sinks.transformer import Transformer
 
 
-def with_attr(obj, attr_name, attr_value):
-    setattr(obj, attr_name, attr_value)
-    return obj
-
-
-class MailTransformer(Transformer):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.file_blocks: List[FileBlock] = []
-
-    def block_to_html(self, block: BaseBlock) -> str:
-        if isinstance(block, FileBlock):
-            self.file_blocks.append(block)
-            return f"<p>See attachment {block.filename}</p>"
-        elif isinstance(block, LinksBlock):
-            if getattr(block, "html_class", None):
-                class_part = f' class="{block.html_class}"'
-            else:
-                class_part = ""
-            return (
-                f"<ul{class_part}>\n"
-                + "\n".join(f'  <li><a href="{link.url}">{link.text}</a></li>' for link in block.links)
-                + "\n</ul>\n"
-            )
-        else:
-            return super().block_to_html(block)
-
-
-class MailSender:
+class MailSender(HTMLBaseSender):
     def __init__(self, mailto: str, account_id: str, cluster_name: str, signing_key: str):
         self.mailto = mailto
         self.signing_key = signing_key
         self.account_id = account_id
         self.cluster_name = cluster_name
 
-    def send_finding_via_email(self, finding: Finding, platform_enabled: bool):
+    def send_finding(self, finding: Finding, platform_enabled: bool):
         blocks: List[BaseBlock] = []
 
         status: FindingStatus = (
@@ -80,7 +52,7 @@ class MailSender:
                 enrichment.blocks = [Transformer.scanReportBlock_to_fileblock(b) for b in enrichment.blocks]
             blocks.extend(enrichment.blocks)
 
-        transformer = MailTransformer()
+        transformer = HTMLTransformer()
         html_body = self.__build_html(transformer.to_html(blocks).strip())
 
         ap_obj = apprise.Apprise()
@@ -151,7 +123,7 @@ class MailSender:
     def __build_html(self, body):
         return f"""<html>
 <style>
-{self.__get_css()}
+{self.get_css()}
 </style>
 <body>
 
@@ -159,53 +131,4 @@ class MailSender:
 
 </body>
 </html>
-"""
-
-    def __get_css(self):
-        return """
-*, body {
-    font-family: Monaco, Menlo, Consolas, "Courier New", monospace, sans-serif;
-    font-size: 12px;
-}
-.header code {
-    background-color: rgba(29, 28, 29, 0.04);
-    border: 1px solid rgba(29, 28, 29, 0.13);
-    border-radius: 3px;
-    box-sizing: border-box;
-    color: rgb(224, 30, 90);
-    padding-bottom: 1px;
-    padding-left: 3px;
-    padding-right: 3px;
-    padding-top: 2px;
-}
-.header b {
-    display: inline-block;
-    margin-left: 1.5em;
-}
-.header {
-    margin-bottom: 1.5em;
-}
-ul.header_links, ul.header_links li {
-    margin: 0;
-    padding: 0;
-}
-ul.header_links {
-    margin-bottom: 3em;
-}
-ul.header_links li {
-    border: 1px solid rgba(29, 28, 29, 0.3);
-    box-sizing: border-box;
-    border-radius: 4px;
-    color: rgb(29, 28, 29);
-    font-weight: bold;
-    display: inline;
-    padding-bottom: 2px;
-    padding-left: 4px;
-    padding-right: 4px;
-    padding-top: 4px;
-}
-ul.header_links li a {
-    color: #555;
-    text-decoration: none;
-}
 """
